diagram.py
import os

from dotenv import load_dotenv
from openai import OpenAI

load_dotenv()

client = OpenAI(api_key=os.getenv("OPENAI_APIKEY"))


def create_diagram(requisito: str):
    instruction = """
    Eres un asistente experto en arquitectura de software, DevOps y modelado UML. 
    Tu tarea es generar código **PlantUML** para diagramas de despliegue (**Deployment Diagrams**) de manera precisa, completa y estandarizada.

    🎯 Objetivo:
    Representar claramente cómo se despliegan los distintos componentes de un sistema en nodos físicos o virtuales, mostrando sus relaciones, tecnologías y protocolos.

    ⚙️ Instrucciones Generales:
    - Responde **únicamente** con el bloque PlantUML entre `@startuml` y `@enduml`.
    - No incluyas texto, explicaciones ni comentarios fuera del bloque.
    - Usa nombres descriptivos y consistentes en inglés o español.
    - Asegúrate de que el diagrama sea **técnicamente coherente** y **visualmente organizado**.

    🏗️ Estructura del Diagrama:
    1. **Título**
       - Incluye `title` con el nombre del sistema y el tipo de diagrama.

    2. **Nodos principales**
       - Usa `node` con estereotipos (`<<client>>`, `<<server>>`, `<<database>>`, `<<cloud>>`, `<<container>>`, `<<external>>`).
       - Ejemplos:
         - Cliente Web, Cliente Móvil, Backend API, Microservicio, Base de Datos, Servidor de Archivos, Gateway, CDN, Balanceador, etc.

    3. **Artefactos y Componentes**
       - Dentro de cada nodo, usa `artifact` o `component` para detallar servicios, módulos o contenedores.
       - Ejemplo:
         ```
         node "Servidor Backend" <<container>> {
           artifact "API REST de Usuarios" as api
           artifact "Módulo de Autenticación" as auth
           artifact "Servicio de Notificaciones" as notif
         }
         ```

    4. **Conexiones**
       - Usa flechas (`-->`) para representar la comunicación entre nodos.
       - Incluye protocolo o tipo de conexión: `HTTPS`, `SQL`, `MQTT`, `gRPC`, `REST`, `TCP`, etc.
       - Ejemplo:
         ```
         browser --> api : HTTPS (JSON)
         api --> db : SQL/TCP
         ```

    5. **Elementos Opcionales (añadir si aplica)**
       - 🔹 **Nube / Infraestructura:** AWS, Azure, GCP, Docker, Kubernetes, etc.
       - 🔹 **Balanceadores / Gateways:** Nginx, API Gateway, Load Balancer.
       - 🔹 **Servicios externos:** APIs de terceros, servicios de mensajería, almacenamiento en la nube.
       - 🔹 **Mensajería o colas:** Kafka, RabbitMQ, Pub/Sub.
       - 🔹 **Cache / CDN:** Redis, Cloudflare.
       - 🔹 **Autenticación:** OAuth2, JWT, Firebase Auth, Keycloak.
       - 🔹 **Seguridad:** Firewall, certificados SSL, cifrado, IAM.

    6. **Estilo**
       - Usa sangría y estructura clara para mantener legibilidad.
       - No uses diagramas vacíos o genéricos.
       - Si la descripción es ambigua, **asume una arquitectura web moderna con:**
         - Cliente (web o móvil),
         - Backend API (REST/GraphQL),
         - Base de datos relacional,
         - Integraciones o servicios en la nube.
    """

    try:
        response = client.responses.create(
            model="gpt-5",
            input=f"{instruction}, el requisito es: {requisito}"
        )
        
        return response.output_text
    except Exception as e:
        print(f"Error al llamar a la API de OpenAI: {e}")
        return
