requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.0
argcomplete==3.6.2
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cowpy==1.1.5
cycler==0.12.1
Django==5.1.2
djangorestframework==3.15.2
dnspython==2.7.0
fastapi==0.115.2
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
mediapipe==0.10.14
ml_dtypes==0.5.0
numpy==2.1.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
OtroRepo @ https://github.com/Loperaa-Juan/OtroRepo/releases/download/0.1.0/otrorepo-0.1.0-py3-none-any.whl#sha256=9c841773f9260fab5d8a97def33da24c65ddd1a9a91a328ef483a51a6925108e
packaging==24.1
pillow==11.0.0
pipx==1.7.1
platformdirs==4.3.8
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pygame==2.6.1
pymongo==4.10.1
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.3
rsa==4.9
scipy==1.14.1
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
sqlparse==0.5.1
starlette==0.38.6
termcolor==3.1.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
userpath==1.9.2
uv==0.7.21
uvicorn==0.31.1
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.0.4
